--- app/controllers/admin/api/v2/application_controller.rb
@@ -1,11 +1,0 @@
-module Admin
-  module Api
-    module V2
-      class ApplicationController < ActionController::Base
-        protect_from_forgery with: :null_session
-
-        include ErrorResponseGenerateable
-      end
-    end
-  end
-end
