--- app/controllers/admin/api/v2/products_controller.rb
@@ -1,15 +1,0 @@
-module Admin
-  module Api
-    module V2
-      class ProductsController < ApplicationController
-        def index
-          @products = Product.includes(:product_stock).all
-        end
-
-        def show
-          @product = Product.find(params[:id])
-        end
-      end
-    end
-  end
-end
