+++ app/controllers/customer_api/v1/application_controller.rb
@@ -1,0 +1,9 @@
+module CustomerApi
+  module V1
+    class ApplicationController < ActionController::Base
+      protect_from_forgery with: :null_session
+
+      include ErrorResponseGenerateable
+    end
+  end
+end
