+++ app/controllers/customer_api/v1/customer_payment_methods_controller.rb
@@ -1,0 +1,14 @@
+module CustomerApi
+  module V1
+    class CustomerPaymentMethodsController < ApplicationController
+      def index
+        return [] if params[:uid].blank?
+
+        customer = Customer.find_by(uid: params[:uid])
+        return [] if customer.blank?
+
+        @customer_payment_methods = customer.customer_payment_methods
+      end
+    end
+  end
+end
