+++ app/javascript/packs/domains/order/services/index.ts
@@ -1,5 +1,5 @@
 export const fetchResources = async (idToken: string) => {
-  const url = `${process.env.REACT_APP_ORIGIN}/admin/api/v2/orders`
+  const url = `${process.env.REACT_APP_ORIGIN}/customer_api/v1/orders`
   return await fetch(url, {
     headers: {
       Authorization: `Bearer ${idToken}`,
@@ -11,7 +11,7 @@ export const createCancellationRequestResources = async (
   idToken: string,
   orderId: number
 ) => {
-  const url = `${process.env.REACT_APP_ORIGIN}/admin/api/v2/cancellation_orders`
+  const url = `${process.env.REACT_APP_ORIGIN}/customer_api/v1/cancellation_orders`
   const data = {
     order_id: orderId,
   }
