+++ app/javascript/packs/domains/product/services/index.ts
@@ -1,3 +1,3 @@
 export const fetchResources = async () => {
-  return await fetch(`${process.env.REACT_APP_ORIGIN}/admin/api/v2/products`)
+  return await fetch(`${process.env.REACT_APP_ORIGIN}/customer_api/v1/products`)
 }
