+++ app/views/customer_api/v1/customer_payment_methods/index.json.jbuilder
