+++ app/views/customer_api/v1/customers/show.json.jbuilder
