+++ app/views/customer_api/v1/orders/show.json.jbuilder
