+++ app/views/customer_api/v1/products/index.json.jbuilder
