+++ app/views/customer_api/v1/products/show.json.jbuilder
