+++ config/routes.rb
@@ -4,13 +4,6 @@ Rails.application.routes.draw do
       namespace :v1 do
         post "/graphql", to: "graphql#execute"
       end
-      namespace :v2, defaults: { format: :json } do
-        resources :products, only: %i[index show]
-        resources :orders, only: %i[index show create]
-        resources :cancellation_orders, only: %i[create]
-        resources :customers, only: %i[show create]
-        resources :customer_payment_methods, only: %i[index show]
-      end
     end
 
     resources :products
@@ -28,8 +21,18 @@ Rails.application.routes.draw do
     delete '/logout',   to: 'sessions#destroy'
   end
 
+  namespace :customer_api do
+    namespace :v1, defaults: { format: :json } do
+      resources :products, only: %i[index show]
+      resources :orders, only: %i[index show create]
+      resources :cancellation_orders, only: %i[create]
+      resources :customers, only: %i[show create]
+      resources :customer_payment_methods, only: %i[index show]
+    end
+  end
+
   resources :products, only: [:index]
 
-  mount Api => '/'
-  mount GrapeSwaggerRails::Engine => '/v2/docs'
+  # mount Api => '/'
+  # mount GrapeSwaggerRails::Engine => '/v2/docs'
 end
