+++ spec/requests/customer_api/products_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe 'Admin::Api::V2::Products', type: :request do
+RSpec.describe 'CustomerApi::V1::Products', type: :request do
   let!(:product) { create(:product, price: 1000) }
   let(:headers) do
     {
@@ -12,9 +12,9 @@ RSpec.describe 'Admin::Api::V2::Products', type: :request do
     create(:product_stock, product: product, stock: 2)
   end
 
-  describe 'GET /admin/api/v2/products' do
+  describe 'GET /customer_api/v1/products' do
     subject do
-      get '/admin/api/v2/products', params: {}, headers: headers
+      get '/customer_api/v1/products', params: {}, headers: headers
       JSON.parse(response.body)
     end
 
@@ -23,9 +23,9 @@ RSpec.describe 'Admin::Api::V2::Products', type: :request do
     end
   end
 
-  describe 'GET /admin/api/v2/products/:id' do
+  describe 'GET /customer_api/v1/products/:id' do
     subject do
-      get "/admin/api/v2/products/#{product.id}", params: {}, headers: headers
+      get "/customer_api/v1/products/#{product.id}", params: {}, headers: headers
       JSON.parse(response.body)
     end
